# app/Modules/Admin/Controllers/ApiUsageController.php
<?php

namespace App\Modules\Admin\Controllers;

use App\Http\Controllers\Controller;
use App\Modules\Admin\Resources\ApiUsageLogResource;
use App\Modules\Admin\Services\ApiUsageService;
use App\Modules\Shared\Traits\ApiResponse;
use Illuminate\Http\JsonResponse;
use Illuminate\Http\Request;
use OpenApi\Attributes as OA;

class ApiUsageController extends Controller
{
    use ApiResponse;

    public function __construct(
        private readonly ApiUsageService $apiUsageService
    ) {
    }

    #[OA\Get(
        path: '/admin/api-usage',
        summary: 'Ver historial de consumo de The Odds API',
        description: 'Permite al administrador consultar el historial de consumo de créditos, requests usados y requests restantes.',
        security: [['sanctum' => []]],
        tags: ['Admin API Usage'],
        parameters: [
            new OA\Parameter(
                name: 'provider',
                description: 'Proveedor externo',
                in: 'query',
                required: false,
                schema: new OA\Schema(type: 'string'),
                example: 'the_odds_api'
            ),
            new OA\Parameter(
                name: 'sport_key',
                description: 'Filtrar consumo por deporte',
                in: 'query',
                required: false,
                schema: new OA\Schema(type: 'string'),
                example: 'soccer_australia_aleague'
            ),
            new OA\Parameter(
                name: 'date_from',
                description: 'Fecha inicial',
                in: 'query',
                required: false,
                schema: new OA\Schema(type: 'string', format: 'date'),
                example: '2026-05-01'
            ),
            new OA\Parameter(
                name: 'date_to',
                description: 'Fecha final',
                in: 'query',
                required: false,
                schema: new OA\Schema(type: 'string', format: 'date'),
                example: '2026-05-31'
            ),
            new OA\Parameter(
                name: 'per_page',
                description: 'Cantidad de registros por página',
                in: 'query',
                required: false,
                schema: new OA\Schema(type: 'integer'),
                example: 15
            ),
        ],
        responses: [
            new OA\Response(response: 200, description: 'Historial de consumo obtenido correctamente'),
            new OA\Response(response: 401, description: 'No autenticado'),
            new OA\Response(response: 403, description: 'No autorizado'),
        ]
    )]
    public function index(Request $request): JsonResponse
    {
        $logs = $this->apiUsageService->history($request->only([
            'provider',
            'sport_key',
            'date_from',
            'date_to',
            'per_page',
        ]));

        return $this->successResponse(
            ApiUsageLogResource::collection($logs)->response()->getData(true),
            'Historial de consumo obtenido correctamente.'
        );
    }

    #[OA\Get(
        path: '/admin/api-usage/summary',
        summary: 'Ver resumen mensual de consumo de The Odds API',
        description: 'Devuelve resumen mensual de créditos usados, requests realizados, requests restantes y consumo agrupado.',
        security: [['sanctum' => []]],
        tags: ['Admin API Usage'],
        parameters: [
            new OA\Parameter(
                name: 'month',
                description: 'Mes a consultar en formato YYYY-MM. Si no se envía, usa el mes actual.',
                in: 'query',
                required: false,
                schema: new OA\Schema(type: 'string'),
                example: '2026-05'
            ),
        ],
        responses: [
            new OA\Response(response: 200, description: 'Resumen de consumo obtenido correctamente'),
            new OA\Response(response: 401, description: 'No autenticado'),
            new OA\Response(response: 403, description: 'No autorizado'),
        ]
    )]
    public function summary(Request $request): JsonResponse
    {
        return $this->successResponse(
            $this->apiUsageService->monthlySummary($request->query('month')),
            'Resumen de consumo obtenido correctamente.'
        );
    }
}