# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Casa de Apuestas API",
        "description": "API para casa de apuestas: autenticación, usuarios, deportes, eventos, cuotas, wallet y apuestas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API versión 1"
        }
    ],
    "paths": {
        "/admin/api-usage": {
            "get": {
                "tags": [
                    "Admin API Usage"
                ],
                "summary": "Ver historial de consumo de The Odds API",
                "description": "Permite al administrador consultar el historial de consumo de créditos, requests usados y requests restantes.",
                "operationId": "2a97f2ba01214f954a3e2dc3f9b11cf7",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "query",
                        "description": "Proveedor externo",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "the_odds_api"
                    },
                    {
                        "name": "sport_key",
                        "in": "query",
                        "description": "Filtrar consumo por deporte",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "soccer_australia_aleague"
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Fecha inicial",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        },
                        "example": "2026-05-01"
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Fecha final",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        },
                        "example": "2026-05-31"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de registros por página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 15
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Historial de consumo obtenido correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/api-usage/summary": {
            "get": {
                "tags": [
                    "Admin API Usage"
                ],
                "summary": "Ver resumen mensual de consumo de The Odds API",
                "description": "Devuelve resumen mensual de créditos usados, requests realizados, requests restantes y consumo agrupado.",
                "operationId": "363bb3f2037885c5a8849d5c1622c5b8",
                "parameters": [
                    {
                        "name": "month",
                        "in": "query",
                        "description": "Mes a consultar en formato YYYY-MM. Si no se envía, usa el mes actual.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2026-05"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resumen de consumo obtenido correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/settings": {
            "get": {
                "tags": [
                    "Admin Settings"
                ],
                "summary": "Listar configuraciones del sistema",
                "description": "Permite al administrador consultar región, mercado y frecuencia de sincronización configurados.",
                "operationId": "7fe6af869c93c38783b399599da88a38",
                "responses": {
                    "200": {
                        "description": "Configuraciones obtenidas correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin Settings"
                ],
                "summary": "Actualizar configuración de cuotas",
                "description": "Permite al administrador cambiar región por defecto, mercado por defecto e intervalo de sincronización.",
                "operationId": "1fffbb5a6b4e001ce02421e31eeb90b4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "odds.default_region": {
                                        "type": "string",
                                        "example": "au",
                                        "enum": [
                                            "us",
                                            "uk",
                                            "eu",
                                            "au"
                                        ]
                                    },
                                    "odds.default_market": {
                                        "type": "string",
                                        "example": "h2h",
                                        "enum": [
                                            "h2h",
                                            "spreads",
                                            "totals",
                                            "outrights"
                                        ]
                                    },
                                    "odds.sync_interval_seconds": {
                                        "type": "integer",
                                        "example": 120,
                                        "maximum": 3600,
                                        "minimum": 15
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Configuraciones actualizadas correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token personal de acceso. Formato: Authorization: Bearer {token}",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}